+++ NEWS.md
@@ -11,6 +11,7 @@ border-radius: 128px;
 - Features
   - Force update all failed imports admin task.
 - Fixes
+  - Fix moving folders to subfolder orphans folders bug.
   - Fix id does not exist redirect.
 
 ## v1.0.2